requirements.txt
selenium
webdriver_manager
pygame
pyyaml
